Deleted kate_spade_ca_us spider as integrated into kate_spade spider
